Homework - Pandas
Attached are the homework files for Pandas:
Heroes1.ipynb - I converted this file from .py and saved it. This is the Jupyter Notebook file
Heroes.py - This is the .py file - same script
Home Observations and link to Jupyter Notebook - This file contains link to the Jupyter Notebook file (localhost) and observations based on the trends in data.